requirements.txt
Django~=2.2
djangorestframework ~= 3.0
markdown
django-filter
django-debug-toolbar

# For postgresql
psycopg2<2.9

oso

# If you are working with the bleeding edge of django-oso, you can link to the dev project on your local system.
django-oso
#-e ../../oso/languages/python/django-oso/

# For testing
Faker
